// Common/src/Model/GameColor.ts
const brown = '#8B4513';
const WHITE = '#FFFFFF';
const BLACK = '#000000';
const RED = '#FF0000';

/**
 * A GameColor is a Color used in the game to color player pieces
 * and to map players to their penguins
 */
export enum GameColor {
  BROWN = '#8B4513',
  WHITE = '#FFFFFF',
  BLACK = '#000000',
  RED = '#FF0000',
}